enable mpi cluster types in siena07/robmon
Note: don't stop the cluster if it is MPI. It might work if you allow dynamical spawning of MPI tasks, but this is not the usual way to use MPI. In fact, some HPC systems may not allow dynamic spawning of MPI tasks. It is better to launch them all in the beginning when you run your script with mpirun.