hack provided by SNA group
To enable the hack that we spoke of earlier, we use the attached function to replace what’s in the optimised RSienaTest package. This updated function toggles the sienaBayes function to either update two population-level parameters (hierarchical = TRUE) or not (hierarchical = FALSE). This additional functionality is enabled through the command insertsource(“sienaBayes.R”)  after loading optimised package.

However, we do not need this function for the test cases, as we do not need to turn off the parameter update.